requirements.txt
asn1crypto==0.23.0
Babel==2.5.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.3
Django==1.11.7
django-phonenumber-field==1.3.0
django-twilio==0.8.0
djangorestframework==3.7.3
enum34==1.1.6
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
phonenumberslite==8.8.5
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.3.0
pytz==2017.3
requests==2.18.4
six==1.11.0
twilio==5.6.0
urllib3==1.22
